+++ features/steps/define_steps.rb
@@ -1,13 +1,13 @@
 class Spinach::Features::Define < Spinach::FeatureSteps
   step 'I run `lyracyst define test`' do
-    pending 'step not implemented'
+    @output = `lyracyst define test`
   end
 
   step 'the output should contain a definition' do
-    pending 'step not implemented'
+    @output =~ /\[Definition\] [a-z-]* - [A-Za-z0-9 ,;:'"\.\(\)”-]*/
   end
 
   step 'the exit status should be 0' do
-    pending 'step not implemented'
+    $?.exitstatus == 0
   end
 end