+++ features/steps/example_steps.rb
@@ -1,13 +1,13 @@
 class Spinach::Features::Example < Spinach::FeatureSteps
   step 'I run `lyracyst example test`' do
-    pending 'step not implemented'
+    @output = `lyracyst example test`
   end
 
   step 'the output should contain an example' do
-    pending 'step not implemented'
+    @output =~ /\[Example\]  - [A-Za-z0-9 ,:\-\='"+;\(\)\/\\.?&#]*/
   end
 
   step 'the exit status should be 0' do
-    pending 'step not implemented'
+    $?.exitstatus == 0
   end
 end