+++ features/steps/info_steps.rb
@@ -1,13 +1,13 @@
 class Spinach::Features::Info < Spinach::FeatureSteps
-  step 'I run `lyracyst info test`' do
-    pending 'step not implemented'
+  step 'I run `lyracyst info fuck`' do
+    @output = `lyracyst info fuck`
   end
 
   step 'the output should contain word information' do
-    pending 'step not implemented'
+    @output =~ /\[Word info\] - [A-Za-z0-9: |ˈfʌk\.]*/
   end
 
   step 'the exit status should be 0' do
-    pending 'step not implemented'
+    $?.exitstatus == 0
   end
 end