+++ features/steps/phrase_steps.rb
@@ -1,13 +1,13 @@
 class Spinach::Features::Phrase < Spinach::FeatureSteps
   step 'I run `lyracyst phrase test`' do
-    pending 'step not implemented'
+    @output = `lyracyst phrase test`
   end
 
   step 'the output should contain a phrase' do
-    pending 'step not implemented'
+    @output =~ /\[Bi-gram phrases\] - [a-z |']*/
   end
 
   step 'the exit status should be 0' do
-    pending 'step not implemented'
+    $?.exitstatus == 0
   end
 end