+++ features/steps/relate_steps.rb
@@ -1,13 +1,13 @@
 class Spinach::Features::Relate < Spinach::FeatureSteps
   step 'I run `lyracyst relate test`' do
-    pending 'step not implemented'
+    @output = `lyracyst relate test`
   end
 
   step 'the output should contain related words' do
-    pending 'step not implemented'
+    @output =~ /\[Related words\] [A-Za-z0-9 \-,'ö]*/
   end
 
   step 'the exit status should be 0' do
-    pending 'step not implemented'
+    $?.exitstatus == 0
   end
 end