+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Changelog
 ===
 
+Version 0.0.8 - Major Improvements
+- The biggest milestone yet!
+- Most Wordnik search functions now work
+- Removed Altervista related words in favor of using Wordnik
+- Migrated from [ARPABET](http://arpabet.heroku.com) to [Rhymebrain](http://rhymebrain.com/api.html) and using it to fetch rhymes, word info, and portmanteaus
+- Migrated from [commander](http://github.com/visionmedia/commander) to [gli](http://github.com/davetron5000/gli) for CLI interface
+
 Version 0.0.7 - Tests, Optimization, & More Cleanup
 - Better docs
 - Testing on more Ruby implementations